Rename to DDR Showdown, Improvements, and more

+++ src/main/java/com/lovetropics/minigames/common/content/escape_race/ddr_machine/levels/DdrLevel.java
@@ -73,8 +73,8 @@ public record DdrLevel(
 		return track.value().lengthInTicks();
 	}
 
-	public long lastRecordedTick() {
-		return inputs.stream().mapToLong(TimedDdrInput::tick).max().orElse(this.lengthInTicks());
+	public int lastRecordedTick() {
+		return Math.toIntExact(inputs.stream().mapToLong(TimedDdrInput::tick).max().orElse(this.lengthInTicks()));
 	}
 
 	public static Path pathFor(Identifier id) {
